- first py file commit
This is the game I work on at the moment. I don't have much time to spend on it so the development goes quite slow. The current problem I work on atm is that I need to add 4 diagonal directions for the aimer of the tank. There might be other minor problems.